requirements.txt
# 教学文档智能生成系统 - 依赖列表
# 生成日期：2026-03-25

# GUI框架
customtkinter>=5.2.0
ttkbootstrap>=1.10.0

# 文档生成
python-docx>=1.1.0
docxtpl>=0.16.0

# HTTP请求
requests>=2.31.0

# PDF转换（可选）
docx2pdf>=0.1.8; sys_platform == 'win32'

# 工具库
pathlib2>=2.3.7